Moved files to core to enable reuse + use them
