+++ src/Umbraco.Infrastructure/Services/Notifications/ContentNotificationExtensions.cs
@@ -1,6 +1,7 @@
 // Copyright (c) Umbraco.
 // See LICENSE for more details.
 
+using Umbraco.Cms.Core.Events;
 using Umbraco.Cms.Core.Models;
 
 namespace Umbraco.Cms.Infrastructure.Services.Notifications
