+++ src/Umbraco.Web.Common/DependencyInjection/UmbracoBuilderExtensions.cs
@@ -55,8 +55,6 @@ using Umbraco.Cms.Web.Common.Routing;
 using Umbraco.Cms.Web.Common.Security;
 using Umbraco.Cms.Web.Common.Templates;
 using Umbraco.Cms.Web.Common.UmbracoContext;
-using Umbraco.Core.Events;
-using static Umbraco.Cms.Core.Cache.HttpContextRequestAppCache;
 using IHostingEnvironment = Umbraco.Cms.Core.Hosting.IHostingEnvironment;
 
 namespace Umbraco.Extensions
